+++ pbge/scenes/terrain.py
@@ -236,34 +236,47 @@ class VariableTerrain(Terrain):
 class AnimTerrain(Terrain):
     frames = (0, 1, 2, 3, 4, 5, 6, 7)
     anim_delay = 4
+    position_dependent = True
 
     @classmethod
     def render_top(self, dest, view, x, y):
         """Draw terrain that should appear in front of a model in the same tile"""
         if self.image_top:
             spr = view.get_terrain_sprite(self.image_top, (x, y), transparent=self.transparent)
-            spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            if self.position_dependent:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            else:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay) % len(self.frames)])
 
     @classmethod
     def render_middle(self, dest, view, x, y):
         """Draw terrain that should appear in front of a model in the same tile"""
         if self.image_middle:
             spr = view.get_terrain_sprite(self.image_middle, (x, y), transparent=self.transparent)
-            spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            if self.position_dependent:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            else:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay) % len(self.frames)])
 
     @classmethod
     def render_biddle(self, dest, view, x, y):
         """Draw terrain that should appear in front of a model under the same tile"""
         if self.image_biddle:
             spr = view.get_terrain_sprite(self.image_biddle, (x, y), transparent=self.transparent)
-            spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            if self.position_dependent:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            else:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay) % len(self.frames)])
 
     @classmethod
     def render_bottom(self, dest, view, x, y):
         """Draw terrain that should appear behind a model in the same tile"""
         if self.image_bottom:
             spr = view.get_terrain_sprite(self.image_bottom, (x, y), transparent=self.transparent)
-            spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            if self.position_dependent:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay + (x + y) * 4) % len(self.frames)])
+            else:
+                spr.render(dest, self.frames[(view.phase // self.anim_delay) % len(self.frames)])
 
 
 class WallTerrain(Terrain):